+++ packages/store/src/services/backend-system.ts
@@ -86,6 +86,13 @@ export class SystemService implements Service<BackendSystem, BackendSystemKey> {
         return this.dataProvider.delete(entity);
     }
 
+    /**
+     * Retrieves all backend systems from the store. Can be filtered by providing retrieval options.
+     * N.B. if no `connectionType` is provided in the options, only systems with `abap_catalog` connection type will be returned.
+     *
+     * @param options - backend service retrieval options
+     * @returns - list of backend systems
+     */
     public async getAll(options?: BackendServiceRetrievalOptions): Promise<BackendSystem[] | []> {
         return this.dataProvider.getAll(options);
     }