+++ packages/ui5-proxy-middleware/src/index.ts
@@ -1,1 +1,2 @@
 export { ui5Proxy, directLoadProxy } from './base';
+export type { ProxyConfig } from './base/types';