+++ packages/jest-environment-ui5/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 5.3.17
+
+### Patch Changes
+
+- 88c68c9: fix(deps): update dependency @ui5/cli to v4.0.44
+
 ## 5.3.16
 
 ### Patch Changes