+++ packages/ui5-info/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-info
 
+## 0.13.13
+
+### Patch Changes
+
+- bb310dc: fix(deps): update dependency semver to v7.7.4
+
 ## 0.13.12
 
 ### Patch Changes